Projeto - Tela de Login/requirements.txt
streamlit==1.27.2
streamlit_authenticator==0.2.3
psycopg2==2.9.9
python-dotenv==1.0.0